Compile Shaders in Zig
Added a new function in `build.zig` that creates a new build step to
compile all the shaders listed in the `SHADERS` constant in `build.zig`
with the Sokol shader compiler. The Sokol shader compiler binary is
chosen depending on the target platform. Anyways, thanks to this, the
Powershell script to compile shaders is no longer necessary. I updated
the Sokol dependency as well because why not?