No More Scripts
Got rid of all the Powershell scripts which were used to compile all
shaders and fetch all required dependencies. I moved over all the stuff
that the scripts did to the `build.zig` file instead which means it's
more streamlined because Powershell scripts can only run in Powershell
of course, meaning that the user was required to have Windows or
something. Now that it's a Zig build step, the user only needs Zig to do
whatever like fetching and updating all dependencies which is far more
convenient and practical. Also made sure to use asserts to stop the
minimum and maximum FOV from having a value that causes the program to
crash and moved the near and far values to the camera since that made
more sense rather than having it hardcoded in the projection matrix.